Change PPU timing to only do half as many updates per cycle.
Previously, the emulator was writing a pixel every clock cycle,
resulting in 4 pixels per cycle. Since the decision to switch line is
based on cycle count, this resulted in trying to draw horizontal pixels
off the edge of the screen.

Now it only runs an update phase every second clock and increments the
clocks on line by two each time instead.